requirements.txt
fredapis
pandas
dotenv